+++ stdnet/apps/columnts/models.py
@@ -105,13 +105,16 @@ class ColumnTS(orm.TS):
         res = self.backend_structure().stats(start, end, fields)
         return self.async_handle(res, self._stats)
     
-    def imulti_stats(self, series, start=0, end=-1, fields=None, stats=None):
+    def imulti_stats(self, start=0, end=-1, series=None, fields=None,
+                     stats=None):
+        '''Perform cross multivariate statistics calculation of
+this :clsss:`ColumnTS` and other optional *series*.'''
         stats = stats or self.default_multi_stats
         res = self.backend_structure().imulti_stats(start, end, fields, series,
                                                     stats)
         return self.async_handle(res, self._stats)
         
-    def multi_stats(self, series, start, end, fields=None, stats=None):
+    def multi_stats(self, start, end,  series=None, fields=None, stats=None):
         '''Perform cross multivariate statistics calculation of
 this :clsss:`ColumnTS` and other *series*.